py/socketio_server/requirements.txt
python-socketio
eventlet

